Refactored and updated imports
