+++ readabilipy/simplifiers/__init__.py
@@ -1,0 +1,9 @@
+from .text import normalise_text, normalise_unicode, normalise_whitespace, strip_control_characters, strip_html_whitespace
+
+__all__ = [
+    "normalise_text",
+    "normalise_unicode",
+    "normalise_whitespace",
+    "strip_control_characters",
+    "strip_html_whitespace",
+]
