+++ tests/checks.py
@@ -1,18 +1,19 @@
 import json
 import os
-from ..readabilipy import parse_to_json, text_manipulation
-from ..readabilipy.json_parser import extract_text_blocks_as_plain_text
+from ..readabilipy import simple_json_from_html_string
+from ..readabilipy.simplifiers import strip_html_whitespace
+from ..readabilipy.simple_json import extract_text_blocks_as_plain_text
 
 
 def check_exact_html_output(test_fragment, expected_output=None):
     """Check that expected output is present when parsing HTML fragment."""
     if expected_output is None:
         expected_output = test_fragment
-    article_json = parse_to_json(test_fragment)
+    article_json = simple_json_from_html_string(test_fragment)
     content = str(article_json["plain_content"])
     # Check that expected output is present after simplifying the HTML
-    normalised_expectation = text_manipulation.simplify_html(expected_output)
-    normalised_result = text_manipulation.simplify_html(content)
+    normalised_expectation = strip_html_whitespace(expected_output)
+    normalised_result = strip_html_whitespace(content)
     print("expectation:", normalised_expectation)
     print("result:", normalised_result)
     assert normalised_expectation == normalised_result
@@ -22,11 +23,11 @@ def check_html_output_contains_text(test_fragment, expected_output=None):
     """Check that expected output is present when parsing HTML fragment."""
     if expected_output is None:
         expected_output = test_fragment
-    article_json = parse_to_json(test_fragment)
+    article_json = simple_json_from_html_string(test_fragment)
     content = str(article_json["plain_content"])
     # Check that expected output is present after simplifying the HTML
-    normalised_expectation = text_manipulation.simplify_html(expected_output)
-    normalised_result = text_manipulation.simplify_html(content)
+    normalised_expectation = strip_html_whitespace(expected_output)
+    normalised_result = strip_html_whitespace(content)
     assert normalised_expectation in normalised_result
 
 
@@ -40,9 +41,9 @@ def check_extract_article(test_filename, expected_filename, content_digests=Fals
 
     # Extract simplified article HTML
     if use_readability_js:
-        article_json = parse_to_json(html, content_digests, node_indexes, use_readability=True)
+        article_json = simple_json_from_html_string(html, content_digests, node_indexes, use_readability=True)
     else:
-        article_json = parse_to_json(html, content_digests, node_indexes)
+        article_json = simple_json_from_html_string(html, content_digests, node_indexes)
 
     # Get expected simplified article HTML
     expected_filepath = os.path.join(os.path.dirname(__file__), test_data_dir, expected_filename)
@@ -75,14 +76,14 @@ def check_extract_paragraphs_as_plain_text(test_filename, expected_filename):
 
 def check_html_has_no_output(test_fragment):
     """Check that no output is present when parsing HTML fragment."""
-    article_json = parse_to_json(test_fragment)
+    article_json = simple_json_from_html_string(test_fragment)
     # Check that there is no output
     assert article_json["plain_content"] is None or article_json["plain_content"] == "<div></div>"
 
 
 def check_html_output_does_not_contain_tag(test_fragment, vetoed_tag):
     """Check that vetoed tag is not present when parsing HTML fragment."""
-    article_json = parse_to_json(test_fragment)
+    article_json = simple_json_from_html_string(test_fragment)
     # Check that neither <tag> nor </tag> appear in the output
     content = str(article_json["plain_content"])
     if content is not None:
