+++ tests/test_benchmarking.py
@@ -1,24 +1,20 @@
 import os
-from ..readabilipy.json_parser import parse_to_json
-from ..readabilipy.extractors.extract_title import extract_title
-from ..readabilipy.extractors.extract_date import extract_date
+from ..readabilipy import simple_json_from_html_string
+from ..readabilipy.extractors import extract_date, extract_title
 
 
-test_filepath = os.path.join(os.path.dirname(__file__), "data", "benchmarkinghuge.html")
-with open(test_filepath) as h:
-    html = h.read()
+TEST_FILEPATH = os.path.join(os.path.dirname(__file__), "data", "benchmarkinghuge.html")
+with open(TEST_FILEPATH) as h:
+    HTML = h.read()
 
 
-def test_benchmark_parse_to_json(benchmark):
-
-    benchmark(parse_to_json, html=html)
+def test_benchmark_simple_json_from_html_string(benchmark):
+    benchmark(simple_json_from_html_string, html=HTML)
 
 
 def test_benchmark_extract_title(benchmark):
-
-    benchmark(extract_title, html=html)
+    benchmark(extract_title, html=HTML)
 
 
 def test_benchmark_extract_date(benchmark):
-
-    benchmark(extract_date, html=html)
+    benchmark(extract_date, html=HTML)
