+++ tests/test_simplifiers_html.py
@@ -1,10 +1,9 @@
 """Tests for plain_html functions."""
 from bs4 import BeautifulSoup
-from ..readabilipy import plain_html, text_manipulation
-
+from ..readabilipy.simplifiers import html
 
 def test_remove_metadata():
-    html = """
+    HTML = """
         <!DOCTYPE html>
         <html>
         <head></head>
@@ -13,13 +12,13 @@ def test_remove_metadata():
         </body>
         </html>
     """
-    soup = BeautifulSoup(html, "html5lib")
-    plain_html.remove_metadata(soup)
+    soup = BeautifulSoup(HTML, "html5lib")
+    html.remove_metadata(soup)
     assert "<!-- Comment here -->" not in str(soup)
 
 
 def test_remove_blacklist():
-    html = """
+    HTML = """
         <html>
         <body>
             <button type="button">Click Me!</button>
@@ -27,31 +26,8 @@ def test_remove_blacklist():
         <body>
         </html>
     """
-    soup = BeautifulSoup(html, "html5lib")
-    plain_html.remove_blacklist(soup)
+    soup = BeautifulSoup(HTML, "html5lib")
+    html.remove_blacklist(soup)
     assert "button" not in str(soup)
 
 
-def test_remove_cdata():
-    """Test all possible methods of CData inclusion. Note that in the final
-    example the '//' prefixes have no effect (since we are not in a <script>)
-    tag and so we expect that the first will be displayed (tested in Chrome and
-    Safari)."""
-    html = """
-        <div>
-            <p>Some text <![CDATA[Text inside two tags]]></p>
-            <![CDATA[Text inside one tag]]>
-        </div>
-        <![CDATA[Text outside tags]]>
-        <script type="text/javascript">
-            //<![CDATA[
-            document.write("<");
-            //]]>
-        </script>
-        //<![CDATA[
-            invalid CDATA block
-        //]]>
-    """.strip()
-    parsed_html = str(plain_html.parse_to_tree(html))
-    expected_output = "<div><div><p>Some text</p></div><p>//</p></div>"
-    assert text_manipulation.simplify_html(parsed_html) == expected_output
