+++ tests/test_weird_html.py
@@ -1,7 +1,7 @@
 """Tests for weird HTML input."""
 from pytest import mark
 from .checks import check_exact_html_output
-from ..readabilipy import text_manipulation
+from ..readabilipy.simplifiers.text import matched_punctuation_marks, terminal_punctuation_marks
 
 
 def test_non_printing_control_characters():
@@ -37,29 +37,7 @@ def test_iframe_with_source():
     )
 
 
-# Test whitespace around tags
-@mark.parametrize('terminal_punctuation', text_manipulation.terminal_punctuation_marks)
-def test_ensure_correct_punctuation_joining(terminal_punctuation):
-    """Do not join with ' ' if the following character is a punctuation mark."""
-    check_exact_html_output("""
-        <div>
-            <p>
-                Some text <a href="example.com">like this</a>{0} with punctuation.
-            </p>
-        </div>""".format(terminal_punctuation),
-        """<div><p>Some text like this{0} with punctuation.</p></div>""".format(terminal_punctuation))
 
-
-@mark.parametrize('matched_pair', text_manipulation.matched_punctuation_marks)
-def test_ensure_correct_bracket_quote_joining(matched_pair):
-    """Do not join with ' ' if the following character is a punctuation mark."""
-    check_exact_html_output("""
-        <div>
-            <p>
-                Some text {0}<a href="example.com">like this</a>{1} with punctuation.
-            </p>
-        </div>""".format(*matched_pair),
-        """<div><p>Some text {0}like this{1} with punctuation.</p></div>""".format(*matched_pair))
 
 
 # Test comments inside tags
